chore: move pacman config and repo into pacman module
