+++ archinstall/lib/pacman/repo.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+class Repo(Enum):
+	Multilib = "multilib"
+	Testing = "testing"
+	MultilibTesting = "multilib-testing"
